% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.rql.R
\name{.append.rql}
\alias{.append.rql}
\title{Insert some RQL statements to perform the query. TODO parse RQL properly!}
\usage{
.append.rql(query, target, select, sort, from, limit, locale)
}
\description{
Insert some RQL statements to perform the query. TODO parse RQL properly!
}
\keyword{internal}
