% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{.authToken}
\alias{.authToken}
\title{Constructs the value for the Authorization header}
\usage{
.authToken(username, password)
}
\description{
Constructs the value for the Authorization header
}
\keyword{internal}
