% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.dars.R
\name{.darDTOToDF}
\alias{.darDTOToDF}
\title{Make a list of DataAccessRequest DTOs a data frame (applies to DARs and their amendments)}
\usage{
.darDTOToDF(res)
}
\description{
Make a list of DataAccessRequest DTOs a data frame (applies to DARs and their amendments)
}
\keyword{internal}
