% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.dars.R
\name{.darStatusDTOToDF}
\alias{.darStatusDTOToDF}
\title{Make a list of StatusChange DTOs a data frame (applies to DARs and their amendments)}
\usage{
.darStatusDTOToDF(darId, res)
}
\description{
Make a list of StatusChange DTOs a data frame (applies to DARs and their amendments)
}
\keyword{internal}
