% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{.extractLabel}
\alias{.extractLabel}
\title{Extract label for locale. If not found, fallback to undefined language label (if any).}
\usage{
.extractLabel(locale = "en", labels = list(), localeKey = "lang",
  valueKey = "value")
}
\description{
Extract label for locale. If not found, fallback to undefined language label (if any).
}
\keyword{internal}
