% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{.extractLabel2}
\alias{.extractLabel2}
\title{Extract label for locale. If not found, fallback to undefined language label (if any).}
\usage{
.extractLabel2(locale = "en", labels = list(), localeKey = "locale",
  valueKey = "text")
}
\description{
Extract label for locale. If not found, fallback to undefined language label (if any).
}
\keyword{internal}
