% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{.extractMicaSessionId}
\alias{.extractMicaSessionId}
\title{Extract micasid from cookie data frame.}
\usage{
.extractMicaSessionId(cookies)
}
\description{
Extract micasid from cookie data frame.
}
\keyword{internal}
