% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{.get}
\alias{.get}
\title{Issues a GET request to mica for the specified resource}
\usage{
.get(mica, ..., query = list())
}
\description{
Issues a GET request to mica for the specified resource
}
\keyword{internal}
