% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{.post}
\alias{.post}
\title{Issues a POST form request to mica for the specified resource}
\usage{
.post(mica, ..., query = list())
}
\description{
Issues a POST form request to mica for the specified resource
}
\keyword{internal}
