% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.dars.R
\name{mica.dar.amendment.form}
\alias{mica.dar.amendment.form}
\title{Get the data access request amendment form}
\usage{
mica.dar.amendment.form(mica)
}
\arguments{
\item{mica}{A Mica object}
}
\description{
Get the data access request amendment form.
}
\examples{
\dontrun{
m <- mica.login("someuser", "somepassword", "https://mica-demo.obiba.org")
mica.dar.amendment.form(m)
mica.logout(m)
}
}
\seealso{
Other data access requests functions: \code{\link{mica.dar.actions}},
  \code{\link{mica.dar.amendment.history}},
  \code{\link{mica.dar.amendments.history}},
  \code{\link{mica.dar.amendments}},
  \code{\link{mica.dar.amendment}},
  \code{\link{mica.dar.form}},
  \code{\link{mica.dar.history}}, \code{\link{mica.dars}},
  \code{\link{mica.dar}}
}
\concept{data access requests functions}
