% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.dars.R
\name{mica.dar.amendment}
\alias{mica.dar.amendment}
\title{Get a amendment}
\usage{
mica.dar.amendment(mica, pid, id)
}
\arguments{
\item{mica}{A Mica object}

\item{pid}{Data access request identifier}

\item{id}{Amendment identifier}
}
\description{
Get a specific data access request amendment.
}
\examples{
\dontrun{
m <- mica.login("someuser", "somepassword", "https://mica-demo.obiba.org")
mica.dar.amendment(m, "12345", "12345-1")
mica.logout(m)
}
}
\seealso{
Other data access requests functions: 
\code{\link{mica.dar}()},
\code{\link{mica.dar.actions}()},
\code{\link{mica.dar.agreements}()},
\code{\link{mica.dar.amendment.form}()},
\code{\link{mica.dar.amendment.history}()},
\code{\link{mica.dar.amendments}()},
\code{\link{mica.dar.amendments.history}()},
\code{\link{mica.dar.form}()},
\code{\link{mica.dar.history}()},
\code{\link{mica.dar.preliminaries}()},
\code{\link{mica.dar.preliminary}()},
\code{\link{mica.dar.preliminary.history}()},
\code{\link{mica.dars}()}
}
\concept{data access requests functions}
