% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.dars.R
\name{mica.dar.amendments.history}
\alias{mica.dar.amendments.history}
\title{Get amendments history}
\usage{
mica.dar.amendments.history(mica, pid, df = TRUE)
}
\arguments{
\item{mica}{A Mica object}

\item{pid}{Data access request identifier}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the history of the amendments of a specific data access request.
}
\seealso{
Other data access requests functions: \code{\link{mica.dar.actions}},
  \code{\link{mica.dar.amendment.form}},
  \code{\link{mica.dar.amendment.history}},
  \code{\link{mica.dar.amendments}},
  \code{\link{mica.dar.amendment}},
  \code{\link{mica.dar.form}},
  \code{\link{mica.dar.history}}, \code{\link{mica.dars}},
  \code{\link{mica.dar}}
}
\concept{data access requests functions}
