% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.dars.R
\name{mica.dar.preliminaries}
\alias{mica.dar.preliminaries}
\title{Get the data access requests preliminary form}
\usage{
mica.dar.preliminaries(mica, status = NULL, df = TRUE)
}
\arguments{
\item{mica}{A Mica object}

\item{status}{Filter by status}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the data access requests preliminary form.
}
\examples{
\dontrun{
m <- mica.login("someuser", "somepassword", "https://mica-demo.obiba.org")
mica.dar.preliminaries(m)
mica.logout(m)
}
}
\seealso{
Other data access requests functions: 
\code{\link{mica.dar}()},
\code{\link{mica.dar.actions}()},
\code{\link{mica.dar.agreements}()},
\code{\link{mica.dar.amendment}()},
\code{\link{mica.dar.amendment.form}()},
\code{\link{mica.dar.amendment.history}()},
\code{\link{mica.dar.amendments}()},
\code{\link{mica.dar.amendments.history}()},
\code{\link{mica.dar.form}()},
\code{\link{mica.dar.history}()},
\code{\link{mica.dar.preliminary}()},
\code{\link{mica.dar.preliminary.history}()},
\code{\link{mica.dars}()}
}
\concept{data access requests functions}
