% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.dars.R
\name{mica.dar}
\alias{mica.dar}
\title{Get a data access request}
\usage{
mica.dar(mica, id)
}
\arguments{
\item{mica}{A Mica object}

\item{id}{Data access request identifier}
}
\description{
Get a specific data access request.
}
