% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.datasets}
\alias{mica.datasets}
\title{Get the datasets}
\usage{
mica.datasets(mica,
  query = "dataset(fields((acronym,name,studyTable,harmonizationTable)),sort(id),limit(0,100))",
  locale = "en", df = TRUE)
}
\arguments{
\item{mica}{A Mica object}

\item{query}{The search query}

\item{locale}{The language for labels (default is "en")}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the datasets
}
