% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.logout}
\alias{mica.logout}
\title{Close connection with Mica}
\usage{
mica.logout(mica)
}
\arguments{
\item{mica}{A Mica object}
}
\description{
Close connection and release resources of Mica.
}
\examples{
\dontrun{
m <- mica.login("https://mica-demo.obiba.org")
mica.logout(m)
}
}
