% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.networks}
\alias{mica.networks}
\title{Get the networks}
\usage{
mica.networks(mica,
  query = "network(fields((acronym.*,name.*)),sort(id),limit(0,100))",
  locale = "en", df = TRUE)
}
\arguments{
\item{mica}{A Mica object}

\item{query}{The search query}

\item{locale}{The language for labels (default is "en")}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the networks
}
