% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.studies}
\alias{mica.studies}
\title{Get the studies}
\usage{
mica.studies(mica, query = NULL, locale = "en")
}
\arguments{
\item{mica}{A Mica object}

\item{query}{The search query}

\item{locale}{The language for labels (default is "en")}
}
\description{
Get the studies
}
