% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.taxonomies}
\alias{mica.taxonomies}
\title{Get the taxonomies}
\usage{
mica.taxonomies(mica, query = NULL, locale = "en", target = "variable")
}
\arguments{
\item{mica}{A Mica object}

\item{query}{The search query}

\item{locale}{The language for labels (default is "en")}

\item{target}{What the taxonomy is about: variable (default), dataset, study, network}
}
\description{
Get the taxonomies
}
