% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.variables}
\alias{mica.variables}
\title{Get the variables}
\usage{
mica.variables(mica,
  query = "variable(fields((attributes.label)),sort(id),limit(0,10000))",
  locale = "en", df = TRUE)
}
\arguments{
\item{mica}{A Mica object}

\item{query}{The search query}

\item{locale}{The language for labels (default is "en")}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the variables
}
