% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_boxplot.R
\name{micar_boxplot}
\alias{micar_boxplot}
\title{Boxplot}
\usage{
micar_boxplot(data, x, y, fill = NULL, title = NULL)
}
\arguments{
\item{data}{Data frame.}

\item{x, y}{Variables for x and y axes.}

\item{fill}{Optional variable for fill color.}

\item{title}{Optional plot title.}
}
\value{
A ggplot2 object.
}
\description{
Creates a boxplot.
}
