% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_grouped_barplot.R
\name{micar_grouped_barplot}
\alias{micar_grouped_barplot}
\title{Grouped barplot}
\usage{
micar_grouped_barplot(data, x, y, group, title = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A character string for x-axis variable.}

\item{y}{A character string for y-axis variable.}

\item{group}{A character string for grouping variable (fill aesthetic).}

\item{title}{Plot title}
}
\value{
A ggplot2 object.
}
\description{
Creates a grouped barplot with bars grouped by a categorical variable.
}
