% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_legend.R
\name{micar_legend}
\alias{micar_legend}
\title{Custom Legend Theme}
\usage{
micar_legend(position = "right", title = "legend")
}
\arguments{
\item{position}{Legend position.}

\item{title}{Title of the legend.}
}
\value{
A ggplot2 theme modification.
}
\description{
Custom Legend Theme
}
