% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_palette.R
\name{micar_palette}
\alias{micar_palette}
\title{micar_palette}
\usage{
micar_palette(name = "blue")
}
\arguments{
\item{name}{Name of the color theme: one of "blue", "red", "purple", "gray".}
}
\value{
A character vector of 3 colors: low, mid, high for gradient scales.
}
\description{
Returns a named color palette used in micaR plots.
}
