% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_stacked_barplot.R
\name{micar_stacked_barplot}
\alias{micar_stacked_barplot}
\title{Stacked barplot}
\usage{
micar_stacked_barplot(data, x, y, group, title = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A character string for x-axis variable.}

\item{y}{A character string for y-axis variable.}

\item{group}{A character string for grouping variable (fill aesthetic).}

\item{title}{Plot title}
}
\value{
A ggplot2 object.
}
\description{
Creates a stacked barplot from a data frame using ggplot2.
}
