% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_plot.R
\name{plsda_plot}
\alias{plsda_plot}
\title{plsda_plot}
\usage{
plsda_plot(data, group, ncomp = 2)
}
\arguments{
\item{data}{A numeric data frame or matrix (features).}

\item{group}{A factor or character vector indicating class.}

\item{ncomp}{Number of components to extract. Default 2.}
}
\value{
A ggplot2 plot.
}
\description{
Perform PLS-DA and plot the result (2D scatter with ellipses).
}
