% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.rql.R
\name{rqlNode}
\alias{rqlNode}
\title{New RQL node}
\usage{
rqlNode(name, ...)
}
\arguments{
\item{name}{The name of the operator}

\item{...}{The arguments of the operation}
}
\description{
Creates a new RQL node that represents an operation.
}
