% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_micar.R
\name{scale_color_micar}
\alias{scale_color_micar}
\title{Discrete color scale using micar palettes}
\usage{
scale_color_micar(palette = "nature", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name (e.g., "nature")}

\item{reverse}{Reverse the color order}

\item{...}{Other arguments to scale_color_manual}
}
\description{
Discrete color scale using micar palettes
}
