% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_micar_bw.R
\name{scale_color_micar_bw}
\alias{scale_color_micar_bw}
\title{Black-and-White ggplot2 Color Scale for micar Palette}
\usage{
scale_color_micar_bw(discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{discrete}{Logical. Use discrete scale? Default TRUE.}

\item{reverse}{Logical. Reverse the palette? Default FALSE.}

\item{...}{Additional parameters passed to `scale_color_manual` or `scale_color_gradientn`.}
}
\value{
A ggplot2 color scale.
}
\description{
Provides a black-and-white (grayscale) color scale for discrete or continuous data.
}
