% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.dars.R
\name{mica.dars}
\alias{mica.dars}
\title{Get the data access requests}
\usage{
mica.dars(mica, df = TRUE)
}
\arguments{
\item{mica}{A Mica object}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the data access requests.
}
