% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.login}
\alias{mica.login}
\title{Mica connection}
\usage{
mica.login(username = getOption("mica.username", "anonymous"),
  password = getOption("mica.password", "password"),
  url = getOption("mica.url"), opts = getOption("mica.opts", list()))
}
\arguments{
\item{username}{User name in mica. Can be provided by "mica.username" option.}

\item{password}{User password in mica. Can be provided by "mica.password" option.}

\item{url}{Mica url or list of mica urls. Can be provided by "mica.url" option.}

\item{opts}{Curl options. Can be provided by "mica.opts" option.}
}
\value{
A Mica object.
}
\description{
Open a connection with Mica.
}
