% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.studies.R
\name{mica.studies}
\alias{mica.studies}
\title{Get the studies}
\usage{
mica.studies(mica,
  query = "study(fields((acronym,name,model.methods.design,populations.dataCollectionEvents.model.dataSources,model.numberOfParticipants.participant)),sort(id),limit(0,100))",
  locale = "en", df = TRUE)
}
\arguments{
\item{mica}{A Mica object}

\item{query}{The search query}

\item{locale}{The language for labels (default is "en")}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the studies
}
