% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.variables}
\alias{mica.variables}
\title{Get the variables}
\usage{
mica.variables(mica, query = NULL)
}
\arguments{
\item{mica}{A Mica object}

\item{query}{The search query}
}
\description{
Get the variables
}
