% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica.R
\name{mica.variables}
\alias{mica.variables}
\title{Get the variables}
\usage{
mica.variables(mica, query = NULL, locale = "en")
}
\arguments{
\item{mica}{A Mica object}

\item{query}{The search query}

\item{locale}{The language for labels (default is "en")}
}
\description{
Get the variables
}
