% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_legend.R
\name{micar_legend}
\alias{micar_legend}
\title{Universal legend beautifier for ggplot2}
\usage{
micar_legend(position = "right", base_size = 12)
}
\arguments{
\item{position}{Position of the legend: "top", "right", "bottom", "left"}

\item{base_size}{Base font size for legend text}
}
\value{
A ggplot2 theme layer
}
\description{
Universal legend beautifier for ggplot2
}
