% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_palette.R
\name{micar_palette}
\alias{micar_palette}
\title{Get a Micar Color Palette}
\usage{
micar_palette(palette = "nature", n = NULL, reverse = FALSE)
}
\arguments{
\item{palette}{One of "nature", "cell", or "bw"}

\item{n}{Number of colors to return (optional)}

\item{reverse}{Reverse the color order}
}
\value{
A character vector of HEX color codes
}
\description{
Get a Micar Color Palette
}
