% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_pca.r
\name{micar_pca}
\alias{micar_pca}
\title{PCA Plot}
\usage{
micar_pca(data, group_var = NULL, scale = TRUE, ellipse = TRUE)
}
\arguments{
\item{data}{A data.frame or tibble containing only numeric variables.}

\item{group_var}{Optional. A grouping variable (factor or character vector) of the same length as the number of rows in \code{data}.}

\item{scale}{Logical. Should the variables be scaled? Default is \code{TRUE}.}

\item{ellipse}{Logical. Should 95% confidence ellipses be drawn for groups? Default is \code{TRUE}.}
}
\value{
A \code{ggplot2} object representing the PCA scatter plot.
}
\description{
Performs Principal Component Analysis (PCA) and returns a ggplot2-based PCA scatter plot.
}
\examples{
iris_numeric <- iris[, 1:4]
group <- iris$Species
micar_pca(iris_numeric, group_var = group)

}
