% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micar_theme.R
\name{micar_theme}
\alias{micar_theme}
\title{A Publication-Ready ggplot2 Theme Inspired by Nature Style}
\usage{
micar_theme(base_size = 12, base_family = "sans")
}
\value{
A ggplot2 theme object.
}
\description{
A Publication-Ready ggplot2 Theme Inspired by Nature Style
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = factor(cyl), y = mpg)) +
  geom_boxplot() +
  micar_theme()
print(p)
}
