% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_micar.R
\name{scale_fill_micar}
\alias{scale_fill_micar}
\title{Discrete fill scale using micar palettes}
\usage{
scale_fill_micar(palette = "nature", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name (e.g., "nature")}

\item{reverse}{Reverse the color order}

\item{...}{Other arguments to scale_fill_manual}
}
\description{
Discrete fill scale using micar palettes
}
